requirements.txt
requests
beautifulsoup4
pandas
python-dotenv
psycopg2-binary
google-genai
pyyaml